Fix: #67- 중복 NSPersistentContainer 생성으로 CoreDataManager 추가
