+++ CurrencyConverter/CurrencyConverter/Data/CoreData/ExchangeRateRecord+CoreDataProperties.swift
