+++ CurrencyConverter/CurrencyConverter/Data/CoreData/FavoriteCurrency+CoreDataClass.swift
