+++ CurrencyConverter/CurrencyConverter/Data/CoreData/FavoriteCurrency+CoreDataProperties.swift
